Move deployment ID derivation function to `airnode-node`
