+++ packages/airnode-node/src/index.ts
@@ -8,5 +8,6 @@ export * from './types';
 export * from './version';
 export * from './config';
 export * from './validation';
+export * from './workers';
 export * from './workers/local-gateways';
 export * from './constants';
